requirements.txt
python-telegram-bot>=21.0
requests
urllib3
beautifulsoup4
cachetools
python-dotenv
dnspython